change generated rust code number type i32 -> f32

+++ src/to_ts_rust/to_rust/common_methods/common_methods.rs
@@ -4,13 +4,13 @@ use crate::type_parser::typed_ast::{TypeFlag, TypedAstType, TypedIdent};
 impl ToRust {
     pub fn type_flag_to_rust(&self, type_flag: TypeFlag) -> String {
         match type_flag {
-            TypeFlag::NumberType => "i32".to_string(),
+            TypeFlag::NumberType => "f32".to_string(),
         }
     }
 
     pub fn typed_ast_type_to_rust(&self, typed_ast_type: TypedAstType) -> String {
         match typed_ast_type {
-            TypedAstType::Number => "i32".to_string(),
+            TypedAstType::Number => "f32".to_string(),
             TypedAstType::Void => "()".to_string(),
             _ => "()".to_string(), // TODO: Funcの型生成するの面倒くさいので後回しにしてます、あとでやりましょう！
         }
